pyproject.toml
[project]
name = "compliance-reasoning-agent"
version = "0.1.0"
description = "基于 LangGraph 的合规推理智能体 - 专利排查、趋势分析、窗口期预警"
requires-python = ">=3.12"
dependencies = [
    # === Agent Framework ===
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-mcp-adapters>=0.1",
    "mcp[cli]>=1.0",

    # === Memory ===
    "mem0ai>=1.0.4",

    # === Backend ===
    "fastapi>=0.133.1",
    "uvicorn[standard]>=0.41.0",

    # === Frontend ===
    "streamlit>=1.40",

    # === Data Processing & Visualization ===
    "pandas>=2.0",
    "plotly>=5.0",

    # === Configuration ===
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",

    # === Patent & Trend APIs ===
    "google-search-results>=2.4.2",
    "pytrends>=4.9",

    # === Mem0 Vector Store (local channel) ===
    "qdrant-client>=1.9",

    # === Database ===
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "redis>=5.0",

    # === Utilities ===
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
